Moved FilePath Class to its own file
